business_system/requirements.txt
psycopg2-binary==2.9.11
APScheduler==3.10.4
Faker==19.6.2
python-dotenv==1.0.0